+++ web/src/reducers/activities.selectors.js
@@ -93,7 +93,8 @@ export const selectActivityCostSummary = createSelector(
         ...o,
         [year]: {
           contractorResources: activity.contractorResources.map(c => ({
-            description: c.name || 'Private Contractor or Vender Name not specified',
+            description:
+              c.name || 'Private Contractor or Vender Name not specified',
             totalCost: c.years[year],
             unitCost: c.hourly.useHourly ? c.hourly.data[year].rate : null,
             units: c.hourly.useHourly