+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Next release
 
-Anticipated release: March 30, 2021
+Anticipated release: April 19, 2021
 
 #### 🚀 New features
 
@@ -17,6 +17,7 @@ Anticipated release: March 30, 2021
 #### ⚙️ Behind the scenes
 
 - Refactor login code ([#2632])
+- Updated field labels and descriptions in the Private Contractor section of activities ([#2940])
 - Add new seed users for different login issues ([#2989])
 
 
@@ -31,4 +32,5 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 [#2830]: https://github.com/CMSgov/eAPD/issues/2830
 [#2864]: https://github.com/CMSgov/eAPD/issues/2864
 [#2938]: https://github.com/CMSgov/eAPD/issues/2938
+[#2940]: https://github.com/CMSgov/eAPD/issues/2940
 [#2989]: https://github.com/CMSgov/eAPD/issues/2989